feat: add ability to merge root module into options
This provides the ability for the root module getters, actions, mutations and modules to be merged into the override options when creating the store.

The default behviour is that the root module will be overridden throwing away the actions/mutations/getters if you set any of these values in the override options.