Fix return type of `dispatch` function (#39)
* Fix return type of `dispatch` function

Vuex checks if the returned type looks like a promise (i.e. has a callable
`then` method) and in that case the value is returned as is. This means that if
I return `A` which is `PromiseLike` (but not an actual promise) the return type
should be `A` not `Promise<A>` as it is currently.

* Add `rootGetters` to `GetterHandlers`